Move array load (via handle) to top level class
